extract EggQuantityCalculator from quantity_calculator to isolate business logic
